Update public class, add block class
I added a block class containing the following properties:
        - String hash
	- String previousHash
	- String data
	- long timeStamp
	- int nonce

The block class includes standard getters and setters for all properties. Block class methods include the mineBlock(), and calculateBlockHash() methods. Will need to complete the buildBlock() and validateBlock() methods.

I included the construction of the genesis block in the main() method. The while loop needs to be completed to include the handling of all lines from the .txt file and their conversion to blocks.