Edits to handling of futures
Made a list to store futures. In the main for loop of the mining method we submit a task, make a future and store the future in the list of futures. If a future is ready we get the value and check. If the future is not ready we keep looping and submitting tasks to the thread pool to execute. This still underperforms, but I haven't tested it beyond a prefix value of 2. Perhaps this method will be more effective as the prefix value increases. We can keep it around for supporting documentation in the final project if we decide to go with another implementation.